New time stepping mechanism
Here I follow Richard's original design, with loops for days, hours, and
minutes. It does limit the time step size, as this has now to give an
integer number of time steps within the hour. But outputting is
simplified in turn. Should be fine, as the time step is of the order of
seconds to tens of seconds.