Temporal interpolation of input data
Add a linear temporal interpolation of input data. This assumes the
input is hourly and adds a simple function for the interpolation. I read
data at both current and next time at the start of each hour. This could
be done slightly more efficiently with pointers, but that's an
optimisation for another time.