Only loop over ocean points
This introduces "ocean_index", an array of {m,n} values for ocean
points. Looping over this and setting m = ocean_index(1,i) and
n = ocean_index(2,i) instead of saying continue when mask(m,n) == 0
should improve the parallelisation.

TODO: The interpolation of input data is still done on all points. It's
not difficult to rewrite this also, but requires quite a bit of code
modifications. This part is also not parallelised, so the gain is
limited.